Add files via upload
A curiosity project that I wrote to polish my c++ skills as well as I got used to the A* algorithm. I created a series of mazes that are written in the files MapCleanBig, MapCleanMedium, and MapCleanMultiple and the idea is that my program finds the goal in the maze (marked with X) from some initial position that the user enters. While A* has as inputs the initial and final position (it knows where the goal is), the other approaches don't know where the goal is so they have to explore all the maze. The mazes are solved by breadth, depth, A*, and a my own customized maze solver based on a variation of what is done in Dijkstra and a greedy search when the goal is found.